explorer.exe ok

+++ modules/disassembler.py
@@ -9,7 +9,7 @@ from modules.constants import BinaryType
 from modules.functions import *
 from modules.instructions import *
 from modules.logger import *
-from modules.utils import get_raw_disassembler
+from modules.utils import get_raw_disassembler, file_exists
 
 PROLOGUES = {
     BinaryType.SCS_32BIT_BINARY.value: [unhexlify("5589E581EC"), unhexlify("5589E55756"), unhexlify("5531C089E5")],
@@ -21,9 +21,10 @@ class EndOfCodeException(Exception):
 
 
 class RecursiveDisassembler(object):
-    def __init__(self, pe):
+    def __init__(self, project):
         super(RecursiveDisassembler, self).__init__()
-        self._pe = pe
+        self._project = project
+        self._pe = project.pe
         self._block_list = BlockList()
         self._function_list = FunctionList()
         self._disassembler = get_raw_disassembler(self._pe.helper.architecture)
@@ -32,16 +33,21 @@ class RecursiveDisassembler(object):
         self._import_addresses = dict()
         self._instructions_cache = dict()
 
-        log_info("Disassembling executable section")
-        for section in self._pe.sections:
-            if section.Characteristics & SECTION_CHARACTERISTICS["IMAGE_SCN_MEM_EXECUTE"]:
-                begin = self._pe.helper.image_base + section.VirtualAddress
-                for i in self._disassembler.disasm(self._pe.helper.mapped[begin:], begin):
-                    self._instructions_cache[i.address] = Instruction(i)
-
-        f = open('entry.pickle', 'wb')
-        pickle.dump(self._instructions_cache, f)
-        f.close()
+        if file_exists(project.binary_path + ".cache"):
+            log_info("Loading disassembled instruction cache")
+            f = open(project.binary_path + ".cache", 'rb')
+            self._instructions_cache = pickle.load(f)
+            f.close()
+        else:
+            log_info("Disassembling executable section")
+            for section in self._pe.sections:
+                if section.Characteristics & SECTION_CHARACTERISTICS["IMAGE_SCN_MEM_EXECUTE"]:
+                    begin = self._pe.helper.image_base + section.VirtualAddress
+                    for i in self._disassembler.disasm(self._pe.helper.mapped[begin:], begin):
+                        self._instructions_cache[i.address] = Instruction(i)
+            f = open(project.binary_path + ".cache", 'wb')
+            pickle.dump(self._instructions_cache, f)
+            f.close()
 
         log_info("Building basic blocks")
         for dll in self._pe.helper.imports:
@@ -68,6 +74,9 @@ class RecursiveDisassembler(object):
         log_info("    Linear recursive disassembly")
         self._remaining_block_disassembly()
 
+        log_info("    Filtering block padding")
+        self._filter_block_padding()
+
         log_info("Function identification")
         self._function_discovery()
 
@@ -79,13 +88,28 @@ class RecursiveDisassembler(object):
     def function_list(self):
         return self._function_list
 
+    def _filter_block_padding(self):
+        for addr in self._block_list.addresses:
+            bb = self._block_list[addr]
+            if bb.starts_with(b"\xCC"):
+                cutSize = 0
+                for instr in bb.instructions:
+                    if instr.bytes == b"\xCC":
+                        cutSize += 1
+                    else:
+                        break
+                if cutSize < bb.size:
+                    new_block = bb.split_at(addr + cutSize)
+                    self._block_list.add(new_block)
+                    self._function_queue.put(new_block.start)
+
     def _function_discovery(self):
         while not self._function_queue.empty():
             addr = self._function_queue.get()
             if addr not in self._function_list:
                 self._function_list.add(Function(self.block_list[addr]))
         count = len(self._function_list)
-        log_info("    using recursive method (%d)" % count)
+        log_info("    using CALL and entrypoints (%d new(s))" % count)
 
         prologues = PROLOGUES[self._pe.helper.architecture]
         block_function = []
@@ -101,7 +125,7 @@ class RecursiveDisassembler(object):
                         break
         for bb in block_function:
             self._function_list.add(Function(bb))
-        log_info("    using prologues method (%d)" % (len(self._function_list) - count))
+        log_info("    using prologues method (%d new(s))" % (len(self._function_list) - count))
 
     def _remaining_block_disassembly(self):
         for section in self._pe.sections:
@@ -114,16 +138,15 @@ class RecursiveDisassembler(object):
                         self._block_queue.put(current_va)
                         try:
                             self._recursive_block_disassembly()
-                            # log_debug("size : %d" % self._block_list[current_va].size)
                             if current_va in self._block_list:
+                                log_debug("size : %d" % self._block_list[current_va].size)
                                 current_va = self._block_list[current_va].start + self._block_list[current_va].size
                             else:
                                 current_va += 1
                         except EndOfCodeException:
-                            print("EndOF " + hex(current_va))
                             current_va = self._block_list.next_after(current_va).start
                     else:
-                        # log_debug("size : %d" % self._block_list[current_va].size)
+                        log_debug("size : %d" % self._block_list[current_va].size)
                         current_va += self._block_list[current_va].size
 
     def _recursive_block_disassembly(self):
@@ -139,14 +162,12 @@ class RecursiveDisassembler(object):
             if begin in self._instructions_cache:
                 i = self._instructions_cache[begin]
             else:
-                rinsrt = list(self._disassembler.disasm(self._pe.helper.mapped[begin:begin + 10], begin))
+                rinsrt = list(self._disassembler.disasm(self._pe.helper.mapped[begin:begin + 16], begin))
                 if len(rinsrt) > 0:
                     i = rinsrt[0]
                 else:
                     i = None
             while i is not None:
-                if i.address == 0x14025a908:
-                    print(5)
                 if i.address != begin and i.address in self._block_list:
                     log_debug("%08x 1-> %08x" % (
                         block.instructions[0].address, self._block_list[i.address].instructions[0].address))
@@ -172,8 +193,9 @@ class RecursiveDisassembler(object):
                         elif i.operands[0].type == X86_OP_IMM:
                             log_debug("call %08x" % i.operands[0].imm)
                             if i.operands[0].imm not in self._import_addresses:
-                                self._block_queue.put(i.operands[0].imm)
-                                self._function_queue.put(i.operands[0].imm)
+                                if i.operands[0].imm >= self._pe.helper.image_base:
+                                    self._block_queue.put(i.operands[0].imm)
+                                    self._function_queue.put(i.operands[0].imm)
                         elif i.operands[0].type == X86_OP_MEM:
                             if i.operands[0].mem.base == 0 and i.operands[0].mem.index == 0 and i.operands[
                                 0].mem.disp != 0:
@@ -194,16 +216,18 @@ class RecursiveDisassembler(object):
                             return block
                         elif i.operands[0].type == X86_OP_MEM:
                             self._block_list.add(block)
-                            if i.operands[0].mem.base == X86_REG_RIP or i.operands[0].mem.base == X86_REG_EIP:
-                                if i.operands[0].mem.disp + i.address + i.size not in self._import_addresses:
-                                    block.link_to(self.__disass_block_from(i.operands[0].mem.disp + i.address + i.size))
+                            if i.operands[0].mem.base != X86_REG_INVALID:
+                                if i.operands[0].mem.base == X86_REG_RIP or i.operands[0].mem.base == X86_REG_EIP:
+                                    if i.operands[0].mem.disp + i.address + i.size not in self._import_addresses:
+                                        block.link_to(
+                                            self.__disass_block_from(i.operands[0].mem.disp + i.address + i.size))
                             else:
                                 if i.operands[0].mem.disp not in self._import_addresses:
                                     block.link_to(self.__disass_block_from(i.operands[0].mem.disp))
                             return block
                         else:
                             raise Exception("JMP case not implemented")
-                    elif i.mnemonic in INSTRUCTIONS_FUNCTION_END:
+                    elif i.mnemonic in INSTRUCTIONS_FUNCTION_END or i.bytes == b"\xcc":
                         self._block_list.add(block)
                         return block
                 if i in self._instructions_cache:
